Dual compile Unix RSA into Algorithms
