+++ src/Common/src/Interop/Unix/System.Security.Cryptography.Native/Interop.ASN1.cs
@@ -6,7 +6,6 @@ using System;
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography;
-using System.Text;
 
 using Microsoft.Win32.SafeHandles;
 
@@ -14,13 +13,6 @@ internal static partial class Interop
 {
     internal static partial class Crypto
     {
-        internal const int NID_undef = 0;
-
-        internal const int NID_X9_62_prime256v1 = 415; // NIST P-256
-        internal const int NID_secp224r1 = 713; // NIST P-224
-        internal const int NID_secp384r1 = 715; // NIST P-384
-        internal const int NID_secp521r1 = 716; // NIST P-521
-
         [DllImport(Libraries.CryptoNative, EntryPoint = "CryptoNative_ObjTxt2Obj", CharSet = CharSet.Ansi)]
         internal static extern SafeAsn1ObjectHandle ObjTxt2Obj(string s);
 
@@ -29,9 +21,6 @@ internal static partial class Interop
 
         [DllImport(Libraries.CryptoNative, EntryPoint = "CryptoNative_GetObjectDefinitionByName", CharSet = CharSet.Ansi)]
         internal static extern IntPtr GetObjectDefinitionByName(string friendlyName);
-
-        [DllImport(Libraries.CryptoNative, EntryPoint = "CryptoNative_ObjSn2Nid", CharSet = CharSet.Ansi)]
-        internal static extern int ObjSn2Nid(string sn);
 
         // Returns shared pointers, should not be tracked as a SafeHandle.
         [DllImport(Libraries.CryptoNative, EntryPoint = "CryptoNative_ObjNid2Obj")]
